feat: seperate flat and nested lists
