+++ src/pages/organizations/pages/organizations-page/index.tsx
@@ -5,7 +5,6 @@ import { RouteComponentProps, withRouter } from 'react-router-dom';
 import { getConfig } from '../../../../config';
 
 import localization from '../../../../lib/localization';
-import { getTranslateText as translate } from '../../../../lib/translateText';
 
 import withOrganizations, {
   Props as OrganizationsProps
@@ -22,8 +21,7 @@ import ErrorPage from '../../../../components/error-page';
 
 import SC from './styled';
 
-import type { OrganizationCatalogSummary } from '../../../../types';
-import { Entity, SortOrder } from '../../../../types/enums';
+import { SortOrder } from '../../../../types/enums';
 import CheckBox from '../../../../components/checkbox';
 import {
   historyPushSearchParams,
@@ -31,6 +29,8 @@ import {
 } from '../../../../lib/location-history-helper';
 
 import CategoryButtons from './category-buttons';
+import OrganizationList from './organization-list';
+import OrganizationCategoriesList from './organization-category-list';
 
 interface Props
   extends OrganizationsProps,
@@ -51,7 +51,7 @@ const OrganizationsPage: FC<Props> = ({
   },
   location,
   history,
-  match: { url }
+  match
 }) => {
   const [searchQuery, setSearchQuery] = useState('');
 
@@ -63,47 +63,7 @@ const OrganizationsPage: FC<Props> = ({
     order: SortOrder.ASC
   });
 
-  const determineNextSortOrder = (nextSelector: string[]) => {
-    const { selector, order } = sortState;
-    const isSameSelector = selector.every(i => nextSelector.includes(i));
-
-    if (!isSameSelector && !nextSelector.includes('prefLabel')) {
-      return SortOrder.DSC;
-    }
-
-    return isSameSelector && order === SortOrder.ASC
-      ? SortOrder.DSC
-      : SortOrder.ASC;
-  };
-
-  const applySort = (selector: string[], nextOrder?: SortOrder) => () => {
-    const order = nextOrder ?? determineNextSortOrder(selector);
-
-    sortOrganizations(selector, order);
-    setSortState({ selector, order });
-  };
-
   const isTransportportal = getConfig().themeNap;
-
-  const filterOrganizationsByName = (query: string) =>
-    query
-      ? organizations.filter(({ prefLabel, name }) =>
-          (translate(prefLabel) ?? name ?? '')
-            .toLowerCase()
-            .includes(query.toLowerCase())
-        )
-      : organizations;
-
-  const renderCaret = (selector: string[]) => {
-    if (sortState.selector.every(i => selector.includes(i))) {
-      return sortState.order === SortOrder.ASC ? (
-        <SC.CaretUp />
-      ) : (
-        <SC.CaretDown />
-      );
-    }
-    return <SC.CaretBoth />;
-  };
 
   const locationSearch = parseSearchParams(location);
   const [includeEmptyOrganizations, setIncludeEmptyOrganizations] = useState(
@@ -136,7 +96,10 @@ const OrganizationsPage: FC<Props> = ({
         );
   }, [includeEmptyOrganizations, currentCategory]);
 
-  return organizations.length > 0 || organizationCategories.length > 0 ? (
+  const entries = () =>
+    currentCategory ? organizationCategories : organizations;
+
+  return entries().length > 0 ? (
     <main id='content' className='container'>
       <div className='row mb-5'>
         <div className='col-12'>
@@ -177,117 +140,26 @@ const OrganizationsPage: FC<Props> = ({
           displayClass='col-12'
         />
       </div>
-      <div className='row'>
-        <SC.SortRow className='col-12'>
-          <SC.Title>
-            <SC.TitleSortButton
-              type='button'
-              onClick={applySort(['prefLabel'])}
-            >
-              {localization.organizationsPage.organization}
-              {renderCaret(['prefLabel'])}
-            </SC.TitleSortButton>
-          </SC.Title>
-          <SC.Info>
-            <SC.SortButton
-              type='button'
-              onClick={applySort(['datasetCount'])}
-              title={localization.organizationsPage.datasetsDescription}
-            >
-              <SC.DatasetIcon />
-              {renderCaret(['datasetCount'])}
-            </SC.SortButton>
-            {!isTransportportal && (
-              <>
-                <SC.SortButton
-                  type='button'
-                  onClick={applySort(['dataserviceCount'])}
-                  title={localization.organizationsPage.dataserviceDescription}
-                >
-                  <SC.ApiIcon />
-                  {renderCaret(['dataserviceCount'])}
-                </SC.SortButton>
-                <SC.SortButton
-                  type='button'
-                  onClick={applySort(['conceptCount'])}
-                  title={localization.organizationsPage.conceptsDescription}
-                >
-                  <SC.ConceptIcon />
-                  {renderCaret(['conceptCount'])}
-                </SC.SortButton>
-                <SC.SortButton
-                  type='button'
-                  onClick={applySort(['informationmodelCount'])}
-                  title={
-                    localization.organizationsPage.informationModelsDescription
-                  }
-                >
-                  <SC.InfomodelIcon />
-                  {renderCaret(['informationmodelCount'])}
-                </SC.SortButton>
-              </>
-            )}
-          </SC.Info>
-        </SC.SortRow>
-        {filterOrganizationsByName(searchQuery).map(
-          (
-            {
-              id,
-              name,
-              prefLabel,
-              datasetCount,
-              dataserviceCount,
-              conceptCount,
-              informationmodelCount
-            },
-            index: number,
-            organizationSummaries: OrganizationCatalogSummary[]
-          ) => {
-            let sortLabel = '';
-            const previousOrganizationName =
-              translate(organizationSummaries[index - 1]?.prefLabel) ||
-              organizationSummaries[index - 1]?.name;
-            const currentOrganizationName = translate(prefLabel) || name;
-
-            if (
-              sortState.selector.includes('prefLabel') &&
-              (index === 0 ||
-                (index > 0 &&
-                  previousOrganizationName &&
-                  currentOrganizationName &&
-                  previousOrganizationName.charAt(0) !==
-                    currentOrganizationName.charAt(0)))
-            ) {
-              sortLabel = currentOrganizationName.charAt(0);
-            }
-
-            return (
-              <SC.Box key={id} className='col-12' to={`${url}/${id}`}>
-                <SC.SortLabel>{sortLabel}</SC.SortLabel>
-                <SC.Title>{translate(prefLabel) || name}</SC.Title>
-                <SC.Info>
-                  <SC.CountTag type={Entity.DATASET}>
-                    {datasetCount}
-                  </SC.CountTag>
-                  {!isTransportportal && (
-                    <>
-                      <SC.CountTag type={Entity.DATA_SERVICE}>
-                        {dataserviceCount}
-                      </SC.CountTag>
-                      <SC.CountTag type={Entity.CONCEPT}>
-                        {conceptCount}
-                      </SC.CountTag>
-                      <SC.CountTag type={Entity.INFORMATION_MODEL}>
-                        {informationmodelCount}
-                      </SC.CountTag>
-                    </>
-                  )}
-                </SC.Info>
-              </SC.Box>
-            );
-          }
-        )}
-      </div>
+      {currentCategory ? (
+        <OrganizationCategoriesList
+          organizationCategories={organizationCategories}
+          searchQuery={searchQuery}
+          history={history}
+          location={location}
+          match={match}
+        />
+      ) : (
+        <OrganizationList
+          organizations={organizations}
+          sortState={sortState}
+          setSortState={setSortState}
+          sort={sortOrganizations}
+          searchQuery={searchQuery}
+          history={history}
+          location={location}
+          match={match}
+        />
+      )}
     </main>
   ) : (
     <Spinner />
